requirements.txt
pytest-django==4.1.0
Django==3.1.4
djangorestframework==3.12.2
psycopg2-binary==2.8.6
PyJWT==1.7.1
djangorestframework-simplejwt==4.4.0
django-cors-headers==3.6.0